Copy test data to user's home directory for Chromium tests
Chromium is a snap, and its confinment rules prevent it from accessing
anything outside of its own snap and the user's home directory.

In particular, this means Chromium cannot access test data stored in
checkbox-kivu snap.

Two jobs are created to copy test data from checkbox-kivu snap to the
user's directory, and to clean it up when testing session is finished,
and Chromium jobs are updated to take the new path into account.